Add Fiscal utility and display fields
These fields mirror the standard calendar fields. Support is only for
fiscal calendars which align on months, so we do not need twice as many
columns. MonthIndex and MTD are both equal for a fiscal and standard
month, for example. We add only columns for display and utility columns
for periods which can differ (mostly year and quarter).

There is a significant slowdown in this version, even for a standard
calendar. Troubleshooting this is necessary before merging these changes
into master.