Add serialized model metadata
Using Tabular Editor's 'Save to folder' feature, I've serialized the
full model to a folder structure. This should allow much better diffing
in the future.

The goal is to make updates using Tabular Editor, and only include the
pbix as a reference file, rather than as the canonical source of truth.
Unfortunately TE does not seem to serialize the M queries.